+++ packages/gamut-labs/src/SkillPathCard/__tests__/SkillPathCard.test.tsx
@@ -1,0 +1,13 @@
+import { setupRtl } from '@codecademy/gamut-tests';
+
+import { SkillPathCard } from '..';
+import { mockSkillPathContent } from './__fixtures__';
+
+const renderView = setupRtl(SkillPathCard);
+
+describe('SkillPathCard', () => {
+  it('renders skill in header', () => {
+    const { view } = renderView(mockSkillPathContent);
+    view.getByText('Skill path');
+  });
+});