+++ packages/gamut-patterns/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.9.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.9.1...@codecademy/gamut-patterns@0.9.2) (2023-02-09)
+
+**Note:** Version bump only for package @codecademy/gamut-patterns
+
+
+
+
+
+### [0.9.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.9.0...@codecademy/gamut-patterns@0.9.1) (2023-02-06)
+
+**Note:** Version bump only for package @codecademy/gamut-patterns
+
+
+
+
+
 ## [0.9.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.8.22...@codecademy/gamut-patterns@0.9.0) (2023-01-19)